major changes
- introduce publication. When the user drops a file, the audio will be created in the db, but not published. When the user clicks publish is when it publishes. Any audio that is not published will be removed through some weekly cron job or something.

+++ server/src/Audiochan.Core/Features/Audios/AudioDetailViewModel.cs
@@ -10,7 +10,7 @@ namespace Audiochan.Core.Features.Audios
         public string Description { get; init; }
         public bool IsPublic { get; init; }
         public string[] Tags { get; init; }
-        public int Duration { get; init; }
+        public decimal Duration { get; init; }
         public long FileSize { get; init; }
         public string FileExt { get; init; }
         public string Picture { get; init; }
